Adds Specific Number of Break Lines in Shiny Ui

The function takes as an input a number and adds that many break lines in the Shiny ui.

n_br(n)

Arguments

n
A non-negative nmeric value

Source

https://stackoverflow.com/questions/46559251/how-to-add-multiple-line-breaks-conveniently-in-shiny

Examples

if (FALSE) {
mainPanel(
plotOutput("plot1"),
n_br(4),
plotOutput("plot3")
)
}